+++ pkg/analysis_server/lib/src/plugin2/generator.dart
@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:analysis_server/src/plugin2/analyzer_version.g.dart';
 import 'package:analyzer/dart/analysis/analysis_options.dart';
 
 /// This class can generate various files to make up the shared plugin package.
@@ -60,6 +61,7 @@ Future<void> main(List<String> args, SendPort sendPort) async {
 name: plugin_entrypoint
 version: 0.0.1
 dependencies:
+  analyzer: '$analyzerVersion'
 ''');
 
     for (var configuration in _pluginConfigurations) {